feat(callsettings): new design for call settings WxC (#246)
