+++ packages/calling/src/Voicemail/constants.ts
@@ -1,21 +1,14 @@
 export const API_V1 = 'api/v1';
 export const VOICEMAIL_FILE = 'VoicemailClient';
 export const BROADWORKS_VOICEMAIL_FILE = 'BroadworksBackendConnector';
-export const BW_XSI_ENDPOINT_VERSION = 'v2.0';
-export const BW_XSI_URL = 'broadworksXsiActionsUrl';
 export const CALLS = 'calls';
 export const BW_TOKEN_FETCH_ENDPOINT = '/idp/bwtoken/fetch';
-export const ENTITLEMENT_BASIC = 'bc-sp-basic';
-export const ENTITLEMENT_BROADWORKS_CONN = 'broadworks-connector';
-export const ENTITLEMENT_STANDARD = 'bc-sp-standard';
 export const JSON_FORMAT = '?format=json';
 export const LIMIT = '&limit';
 export const MARK_AS_READ = 'MarkAsRead';
 export const MARK_AS_UNREAD = 'MarkAsUnread';
 export const MESSAGE_MEDIA_CONTENT = 'messageMediaContent';
 export const MESSAGE_SUMMARY = 'MessageSummary';
-export const NATIVE_SIP_CALL_TO_UCM = 'NATIVE_SIP_CALL_TO_UCM';
-export const NATIVE_WEBEX_TEAMS_CALLING = 'NATIVE_WEBEX_TEAMS_CALLING';
 export const NO_VOICEMAIL_MSG = 'No additional voicemails';
 export const NO_VOICEMAIL_STATUS_CODE = 204;
 export const OFFSET = '?offset';
@@ -27,11 +20,9 @@ export const SORT_ORDER = '&sortOrder';
 export const TRANSCRIPT_STATUS = 'status';
 export const SUMMARY = 'summary';
 export const TRANSCRIPT_CONTENT = 'content';
-export const UCM_VOICEMAIL_FILE = 'UcmBackendConnector';
 export const VMGATEWAY = 'vmgateway';
 export const VOICEMAILS = 'voicemails';
 export const VOICE_MESSAGING_MESSAGES = 'VoiceMessagingMessages';
-export const WEBEX_CALLING_VOICEMAIL_FILE = 'WxCallBackendConnector';
 export const NEW_MESSAGES = 'newMessages';
 export const OLD_MESSAGES = 'oldMessages';
 export const NEW_URGENT_MESSAGES = 'newUrgentMessages';
