+++ packages/calling/src/common/constants.ts
@@ -28,3 +28,12 @@ export const XSI_ACTION_ENDPOINT = 'xsiActionsEndpoint';
 export const XSI_ACTION_ENDPOINT_ORG_URL_PARAM = 'organizations?callingData=true';
 export const SERVICES_ENDPOINT = 'services';
 export const INFER_ID_CONSTANT = 'ciscospark://us';
+export const ENTITLEMENT_BASIC = 'bc-sp-basic';
+export const ENTITLEMENT_BROADWORKS_CONN = 'broadworks-connector';
+export const ENTITLEMENT_STANDARD = 'bc-sp-standard';
+export const NATIVE_SIP_CALL_TO_UCM = 'NATIVE_SIP_CALL_TO_UCM';
+export const NATIVE_WEBEX_TEAMS_CALLING = 'NATIVE_WEBEX_TEAMS_CALLING';
+export const BW_XSI_ENDPOINT_VERSION = 'v2.0';
+export const BW_XSI_URL = 'broadworksXsiActionsUrl';
+export const WEBEX_CALLING_CONNECTOR_FILE = 'WxCallBackendConnector';
+export const UCM_CONNECTOR_FILE = 'UcmBackendConnector';
