+++ packages/calling/src/common/types.ts
@@ -19,6 +19,13 @@ export enum MobiusStatus {
   DEFAULT = 'default',
 }
 
+export enum CALLING_BACKEND {
+  WXC = 'WEBEX_CALLING',
+  BWRKS = 'BROADWORKS_CALLING',
+  UCM = 'UCM_CALLING',
+  INVALID = 'Calling backend is currently not supported',
+}
+
 export type DeviceList = unknown;
 export type CallId = string; // guid;
 export type CorrelationId = string;
